requirements.txt
albumentations==1.1.0
dash==2.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.0
Flask-Compress==1.11
httplib2==0.14.0
bounded-pool-executor==0.0.3
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
imageio==2.16.1
joblib==1.1.0
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pqdm==0.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.3.25
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
gunicorn==20.0.4
